requirements.txt
fastapi
uvicorn
modelscope
transformers
torch
torchaudio
sentencepiece
numpy
ffmpeg-python
scipy
requests
